let's make some more realistic hierarchy and split functions into modules
