+++ .gitignore
@@ -1,1 +1,1 @@
-config.yaml
+__pycache__
