+++ Abominable-Intelligence/config.yaml
@@ -1,0 +1,4 @@
+{
+  'bot_token': bot_token,
+  'guild_id': guild_id
+}
